Move student code out of course package and into student package
