+++ src/main/java/ch/uzh/ifi/access/student/dto/StudentAnswerDTO.java
@@ -1,10 +1,10 @@
-package ch.uzh.ifi.access.course.dto;
+package ch.uzh.ifi.access.student.dto;
 
 import ch.uzh.ifi.access.course.model.ExerciseType;
-import ch.uzh.ifi.access.course.model.workspace.CodeSubmission;
-import ch.uzh.ifi.access.course.model.workspace.MultipleChoiceSubmission;
-import ch.uzh.ifi.access.course.model.workspace.StudentSubmission;
-import ch.uzh.ifi.access.course.model.workspace.TextSubmission;
+import ch.uzh.ifi.access.student.model.CodeSubmission;
+import ch.uzh.ifi.access.student.model.MultipleChoiceSubmission;
+import ch.uzh.ifi.access.student.model.StudentSubmission;
+import ch.uzh.ifi.access.student.model.TextSubmission;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.AllArgsConstructor;
