+++ src/test/java/ch/uzh/ifi/access/student/service/StudentSubmissionServiceTest.java
@@ -1,13 +1,13 @@
-package ch.uzh.ifi.access.course.service;
+package ch.uzh.ifi.access.student.service;
 
 import ch.uzh.ifi.access.TestObjectFactory;
-import ch.uzh.ifi.access.course.dao.StudentSubmissionRepository;
 import ch.uzh.ifi.access.course.model.Assignment;
 import ch.uzh.ifi.access.course.model.Exercise;
-import ch.uzh.ifi.access.course.model.workspace.CodeSubmission;
-import ch.uzh.ifi.access.course.model.workspace.MultipleChoiceSubmission;
-import ch.uzh.ifi.access.course.model.workspace.StudentSubmission;
-import ch.uzh.ifi.access.course.model.workspace.TextSubmission;
+import ch.uzh.ifi.access.student.dao.StudentSubmissionRepository;
+import ch.uzh.ifi.access.student.model.CodeSubmission;
+import ch.uzh.ifi.access.student.model.MultipleChoiceSubmission;
+import ch.uzh.ifi.access.student.model.StudentSubmission;
+import ch.uzh.ifi.access.student.model.TextSubmission;
 import org.assertj.core.api.Assertions;
 import org.junit.After;
 import org.junit.Before;
